+++ src/ProxyDiContainer.ts
@@ -11,17 +11,24 @@ import {
     constructorInjections,
     findInjectableId,
     injectableClasses,
-} from './injectable';
+} from './injectable.decorator';
 import {
     Injection,
     ContainerSettings as ContainerSettings,
     DependencyId,
 } from './types';
 import { DEFAULT_SETTINGS } from './presets';
-import { makeInjectionProxy, makeDependencyProxy } from './Proxy.utils';
-import { makeConstructorDependencyProxy } from './proxy.constuctor';
-import { middlewaresClasses } from './middleware/middleware';
-import { MiddlewareListener } from './middleware/MiddlewareListener';
+import { makeInjectionProxy } from './makeInjectionProxy';
+import { makeDependencyProxy } from './makeDependencyProxy';
+import { makeConstructorDependencyProxy } from './makeConstructorDependencyProxy';
+import { middlewaresClasses } from './middleware/middleware.decorator';
+import { MiddlewareManager } from './middleware/MiddlewaresManager';
+import {
+    MiddlewareContext,
+    MiddlewareRegistrator,
+    MiddlewareRemover,
+    MiddlewareResolver,
+} from './middleware/middleware.api';
 
 /**
  * A dependency injection container
@@ -52,17 +59,15 @@ export class ProxyDiContainer implements IProxyDiContainer {
     /**
      * Holds proxies for dependencies registered in parent containers to provide for it dependencies from this container
      */
-    private parentDependencyProxies: Record<
-        DependencyId,
-        ContainerizedDependency
-    > = {};
+    private inContextProxies: Record<DependencyId, ContainerizedDependency> =
+        {};
 
     /**
      * Settings that control the behavior of the container and it's children
      */
     public readonly settings: Required<ContainerSettings>;
 
-    private middlewareListener: MiddlewareListener;
+    private middlewareManager: MiddlewareManager;
 
     /**
      * Creates a new instance of ProxyDiContainer.
@@ -72,8 +77,8 @@ export class ProxyDiContainer implements IProxyDiContainer {
     constructor(settings?: ContainerSettings, parent?: ProxyDiContainer) {
         this.id = ProxyDiContainer.idCounter++;
 
-        this.middlewareListener = new MiddlewareListener(
-            parent?.middlewareListener
+        this.middlewareManager = new MiddlewareManager(
+            parent?.middlewareManager
         );
 
         if (parent) {
@@ -82,6 +87,24 @@ export class ProxyDiContainer implements IProxyDiContainer {
         }
 
         this.settings = { ...DEFAULT_SETTINGS, ...settings };
+    }
+
+    registerMiddleware(
+        middleware:
+            | MiddlewareRegistrator
+            | MiddlewareRemover
+            | MiddlewareResolver
+    ) {
+        this.middlewareManager.add(middleware);
+    }
+
+    removeMiddleware(
+        middleware:
+            | MiddlewareRegistrator
+            | MiddlewareRemover
+            | MiddlewareResolver
+    ) {
+        this.middlewareManager.remove(middleware);
     }
 
     /**
@@ -147,10 +170,16 @@ export class ProxyDiContainer implements IProxyDiContainer {
 
         const constructorName = instance.constructor?.name;
         if (constructorName && middlewaresClasses[constructorName]) {
-            this.middlewareListener.add(instance);
+            this.middlewareManager.add(instance);
         }
 
-        this.middlewareListener.onRegister(this, id, instance);
+        let context: MiddlewareContext<any> = {
+            container: this,
+            dependencyId: id,
+            dependency: instance,
+        };
+
+        this.middlewareManager.onRegister(context);
 
         return instance;
     }
@@ -176,7 +205,7 @@ export class ProxyDiContainer implements IProxyDiContainer {
      */
     isKnown(dependencyId: DependencyId): boolean {
         return !!(
-            this.parentDependencyProxies[dependencyId] ||
+            this.inContextProxies[dependencyId] ||
             this.dependencies[dependencyId] ||
             (this.parent && this.parent.isKnown(dependencyId)) ||
             injectableClasses[dependencyId]
@@ -213,12 +242,26 @@ export class ProxyDiContainer implements IProxyDiContainer {
             );
         }
 
-        const proxy = this.parentDependencyProxies[dependency];
+        let context: MiddlewareContext<T> = {
+            container: this,
+            dependencyId: dependency,
+            dependency: this.resolveImpl(dependency),
+        };
+
+        context = this.middlewareManager.onResolve(context);
+
+        return context.dependency;
+    }
+
+    private resolveImpl = <T>(
+        dependencyId: DependencyId
+    ): T & ContainerizedDependency => {
+        const proxy = this.inContextProxies[dependencyId];
         if (proxy) {
             return proxy as T & ContainerizedDependency;
         }
 
-        const instance = this.findDependency<T>(dependency);
+        const instance = this.findDependency<T>(dependencyId);
         if (instance) {
             if (
                 instance[PROXYDI_CONTAINER] !== this &&
@@ -227,15 +270,15 @@ export class ProxyDiContainer implements IProxyDiContainer {
             ) {
                 const proxy = makeDependencyProxy(instance);
                 this.injectDependenciesTo(proxy);
-                this.parentDependencyProxies[dependency] = proxy;
-                return proxy as any as T & ContainerizedDependency;
+                this.inContextProxies[dependencyId] = proxy;
+                return proxy as T & ContainerizedDependency;
             }
             return instance;
         }
 
-        const InjectableClass = injectableClasses[dependency];
-        return this.register(InjectableClass, dependency);
-    }
+        const InjectableClass = injectableClasses[dependencyId];
+        return this.register(InjectableClass, dependencyId);
+    };
 
     /**
      * Injects dependencies to the given object based on its defined injections metadata. Does not affect the container.
@@ -308,7 +351,7 @@ export class ProxyDiContainer implements IProxyDiContainer {
         if (dependency) {
             const constructorName = dependency.constructor?.name;
             if (constructorName && middlewaresClasses[constructorName]) {
-                this.middlewareListener.remove(dependency);
+                this.middlewareManager.remove(dependency);
             }
 
             const dependencyInjects: Injections = dependency[INJECTIONS]
@@ -321,7 +364,11 @@ export class ProxyDiContainer implements IProxyDiContainer {
 
             delete this.dependencies[id];
 
-            this.middlewareListener.onRemove(this, id, dependency);
+            this.middlewareManager.onRemove({
+                container: this,
+                dependencyId: id,
+                dependency,
+            });
         }
     }