+++ src/makeInjectionProxy.ts
@@ -6,10 +6,9 @@ import {
     Injection,
     IProxyDiContainer,
     PROXYDI_CONTAINER,
-    IS_INSTANCE_PROXY,
 } from './types';
 
-class InjectionProxy implements IInjectionProxy {
+export class InjectionProxy implements IInjectionProxy {
     [IS_INJECTION_PROXY]: true = true;
     readonly [INJECTION_OWNER]: ContainerizedDependency;
     readonly [PROXYDI_CONTAINER]: IProxyDiContainer;
@@ -68,40 +67,6 @@ export const makeInjectionProxy = <T>(
     }) as T;
 };
 
-export function makeDependencyProxy(dependency: any) {
-    const injectionValues: Record<string | symbol, any> = {};
-
-    return new Proxy(dependency, {
-        get: function (
-            target: InjectionProxy,
-            prop: string | symbol,
-            receiver: any
-        ) {
-            if (prop === IS_INSTANCE_PROXY) {
-                return true;
-            }
-            if (injectionValues[prop]) {
-                return injectionValues[prop];
-            }
-
-            return Reflect.get(target, prop, receiver);
-        },
-
-        set: function (
-            target: InjectionProxy,
-            prop: string | symbol,
-            value: any
-        ) {
-            injectionValues[prop] = value;
-            return Reflect.set(target, prop, value);
-        },
-    });
-}
-
 export function isInjectionProxy(value: any): boolean {
     return !!(value && value[IS_INJECTION_PROXY]);
-}
-
-export function isInstanceProxy(value: any): boolean {
-    return !!(value && value[IS_INSTANCE_PROXY]);
 }